requirements.txt
jinja2
markdown
python-frontmatter
requests
google-genai
pygments